tests/wpt/web-platform-tests/app-history/navigate-event/transitionWhile-same-document-history-back.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
async_test(t => {
  // Wait for after the load event so that the navigation doesn't get converted
  // into a replace navigation.
  window.onload = () => t.step_timeout(() => {
    let onnavigate_calls = 0;
    let onnavigatesuccess_calls = 0;
    appHistory.onnavigate = e => {
      onnavigate_calls++;
      e.transitionWhile(Promise.resolve());
    }
    appHistory.onnavigatesuccess = t.step_func(e => {
      onnavigatesuccess_calls++;
      if (onnavigatesuccess_calls == 3) {
        assert_equals(appHistory.entries().length, 3);
        assert_equals(appHistory.current.index, 1);
        assert_equals(onnavigate_calls, 3);
        history.back();
      } else if (onnavigatesuccess_calls == 4) {
        assert_equals(appHistory.entries().length, 3);
        assert_equals(appHistory.current.index, 0);
        assert_equals(onnavigate_calls, 4);
        t.done();
      }
    });

    appHistory.navigate("?foo").finished
      .then(t.step_func(() => appHistory.navigate("?bar").finished))
      .then(t.step_func(() => {
        assert_equals(appHistory.entries().length, 3);
        assert_equals(appHistory.current.index, 2);
        assert_equals(onnavigate_calls, 2);
        history.back();
      }));
  }, 0);
}, "AppHistoryNavigateEvent's transitionWhile() can intercept same-document history.back()");
</script>
